HOUSING PRICES IN LAGOS
My Group Project Housing prices in Lagos present a compelling case study, and the analysis here provides valuable insights into trends, the average cost of affordable homes, and the most expensive areas in the city. I played a key role in this project by cleaning the data and creating a comprehensive dashboard using Power BI and Excel for effective visualization.